undo waitForTimeToElapse, ensure non-zero instead; rename counter variable
